Apply patch for glibc 2.38 compatibility in bazel build system (#2327)
Based on the commit from
[android.googlesource.com](https://android.googlesource.com/kernel/common/+/75f82c6a15c4188cbb32825892fc6ae3e95479f0%5E%21/),
this change applies a patch to the bazel build system for source code
that does not include this fix and is using glibc version 2.38 or later.
This ensures that the build process does not encounter linkage errors
related to undefined symbols such as __isoc23_strtol, __isoc23_strtoul,
and __isoc23_strtoull.